# Add a `vectorized` option to mesolve