--- a/tests/mesolve/test_adaptive.py
+++ b/tests/mesolve/test_adaptive.py
@@ -1,5 +1,6 @@
 import pytest
 
+from dynamiqs import Options
 from dynamiqs.gradient import BackwardCheckpointed, Direct, Forward
 from dynamiqs.method import Tsit5
 
@@ -13,8 +14,10 @@ from .open_system import dense_ocavity, dia_ocavity, otdqubit
 @pytest.mark.run(order=TEST_LONG)
 class TestMESolveAdaptive(IntegratorTester):
     @pytest.mark.parametrize('system', [dense_ocavity, dia_ocavity, otdqubit])
-    def test_correctness(self, system):
-        self._test_correctness(system, Tsit5())
+    @pytest.mark.parametrize('vectorized', [True, False])
+    def test_correctness(self, system, vectorized):
+        options = Options(vectorized=vectorized)
+        self._test_correctness(system, Tsit5(), options=options)
 
     @pytest.mark.parametrize('system', [dense_ocavity, dia_ocavity, otdqubit])
     @pytest.mark.parametrize('gradient', [Direct(), BackwardCheckpointed(), Forward()])