A rather straightforward PR to implement the `vectorized` option to `mesolve`.

The speedup is not particularly impressive however (non-existant in my benchmark actually):
```text
Vectorized, dense: 1.96 ms ± 69.9 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
Unvectorized, dense: 4.1 ms ± 1.03 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)
Vectorized, dia: 2.01 ms ± 55.1 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
Unvectorized, dia: 1.88 ms ± 20.6 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

It is likely that we need to improve the code of `slindbladian` and underlying functions to get better speed improvements, cc @derekeverett.

<details>
<summary>Benchmark code: time-dependent pi-pulse on a two-level system</summary>

```python
import dynamiqs as dq
import jax.numpy as jnp
from jax.scipy.special import erf
from functools import partial

# set global layout
dq.set_layout('dia')

def pulse(t, T, sigma):
    """Gaussian pulse ansatz."""
    angle = jnp.pi / 2
    norm = jnp.sqrt(2 * jnp.pi) * sigma * T * erf(1 / (2 * jnp.sqrt(2) * sigma))
    gaussian = jnp.exp(-(t - T / 2)**2 / (2 * T**2 * sigma**2))
    return angle * gaussian / norm

# simulation parameters
T1 = 0.01
Tphi = 0.01
nth = 0.1
sigma = 0.2
T = 1.0
ntsave = 11

# time-dependent Hamiltonian
f = partial(pulse, T=T, sigma=sigma)
H = dq.modulated(f, dq.sigmax())

# other mesolve inputs
jump_ops = [
    jnp.sqrt(1 / T1) * dq.sigmam(),
    jnp.sqrt(1 / Tphi) * dq.sigmaz(),
    jnp.sqrt(nth / T1) * dq.sigmap()
]
psi0 = dq.ground()
tsave = jnp.linspace(0.0, T, ntsave)
exp_ops = [dq.ground_dm(), dq.excited_dm()]

# vectorized form
options = dq.Options(vectorized=True, progress_meter=False)
%timeit -n1 -r1 -q dq.mesolve(H, jump_ops, psi0, tsave, exp_ops=exp_ops, options=options)
%timeit dq.mesolve(H, jump_ops, psi0, tsave, exp_ops=exp_ops, options=options)

# unvectorized form
options = dq.Options(vectorized=False, progress_meter=False)
%timeit -n1 -r1 -q dq.mesolve(H, jump_ops, psi0, tsave, exp_ops=exp_ops, options=options)
%timeit dq.mesolve(H, jump_ops, psi0, tsave, exp_ops=exp_ops, options=options)
```

</details>