@pierreguilmin 2025-08-15
@gautierronan, if you try in dense with a strongly-varying drive, e.g. `H = dq.modulated(lambda t: jnp.cos(1e4 * t), dq.sigmax())`, there's a big difference
```
8.5 ms ± 877 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
73.8 ms ± 3.93 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```